cli: Add uvvid command and package
* Add `uvvid` command that will be created using setup.py

* Make `uvvid` look more like a package thanks to a `setup.py` file and
  a clean folder/file structure.

* Move data files into the `data/` directory.
