+++ README.md
@@ -2,3 +2,15 @@ uvvid
 =====
 
 Micro Vector Video
+
+Installation and testing
+------------------------
+
+The `uvvid` package and command can both be installing by running the following
+command from the root of this directory.
+
+    $ pip install --editable .
+
+After installation, you can test its functionality by running
+
+    $ uvvid view --cursor data/cursor.png --video data/exponents.mp4
