+++ data/cursor.png
